Faster EC2 Provisioning
- Override the Strategy for Provisioning for EC2Cloud so that we
immediately call provision if need be when we receive a queued job.
- We also explicitly call for the EC2Cloud's provision method to
provision a machine if we have excess work, which is what it should do.
This may cause unintended consequences when you have a stopped instance
as it may not resume it, but that's not our use case.
- For this code to ever be merged back to master, this should be
optional in the plugin and probably thought about a little deeper.